\name{create_biplot_Color}
\alias{create_biplot_Color}
\title{Create Biplot Color}
\usage{
create_biplot_Color(list, df, x)
}
\arguments{
  \item{list}{list of PCA output}
  \item{df}{dataframe}
  \item{x}{column name}
}
\description{
Prints a ggplot that represents the PCA biplot for a given dataframe with points colored according to a specified variable.
}

