% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaws_lfa.R
\name{jaws.lfa}
\alias{jaws.lfa}
\title{The Jackstraw Weighted Shrinkage Estimation Method for Sparse Coefficients in Logistic Factor Analysis}
\usage{
jaws.lfa(dat, p = NULL, r = NULL, s = NULL, B = NULL, verbose = TRUE,
  seed = NULL, save.all = FALSE)
}
\arguments{
\item{dat}{a data matrix with \code{m} rows as variables and \code{n} columns as observations.}

\item{p}{\code{m*r} matrix of p-values obtained from running \code{jackstraw.LFA}.}

\item{r}{a number (a positive integer) of significance logistic factors.}

\item{s}{a number (a positive integer) of ``synthetic'' null variables (optional).}

\item{B}{a number (a positive integer) of resampling iterations (optional).}

\item{verbose}{a logical specifying to print the progress (default: TRUE).}

\item{seed}{a seed for the random number generator (optional).}

\item{save.all}{a logical specifying to save all objects, including a large SVD object (default: FALSE).}
}
\value{
\code{jaws.lfa} returns a list consisting of
\item{p}{p-values for association tests between variables and each of \code{r} logistic factors}
\item{pi0}{proportion of variables not associated with \code{r} logistic factors, individually}
\item{PIP}{a list of outputs derived from the posterior inclusion probabilities method (including \code{pr}, \code{coefs}, \code{af}, \code{r2})}
\item{PNV}{a list of outputs derived from the proportion of null variables method (including \code{pi0}, \code{coefs}, \code{af}, \code{r2})}
}
\description{
Estimates sparse/shrunken coefficients (or loadings) of Logistic Factor Analysis (LFA).
Based on statistical sginificance of association between variables (i.e., SNPs) and logistic factors (LFs),
the conventional coefficients of LFs are shruken towards zeros, which improve its accuracy.
The only required inputs are the data matrix \code{dat} and the number of LFs \code{r}.
}
\details{
By default, \code{jaws.lfa} computes two canonical jackstraw weighted shrinkage estimators, namely \code{PIP} and \code{PNV}.

It is strongly advised that you take a careful look at your data and use appropriate graphical and statistical criteria
to determine \code{r} (Note that \code{r} does not include an intercept term).

If \code{s} is not supplied, \code{s} is set to about 10\% of \code{m} variables.
If \code{B} is not supplied, \code{B} is set to \code{m*10/s}.

NOTE that you may input a \code{m*r} matrix of p-values obtained from running \code{jackstraw.LFA}.
}
\section{Detailed explanation of the output contained in \code{PIP} and \code{PNV}}{

\describe{
  \item{pr}{a matrix of posterior inclusion probabilities (equivalent to 1-lfdr) for \code{m} coefficients and \code{r} LFs.}
  \item{pi0}{a vector of estimated proportion of null variables for \code{r} LFs.}
  \item{coefs}{a \code{m*r} matrix of shrunken coefficients.}
  \item{af}{a \code{m*n} matrix of shrunken allele frequencies.}
  \item{r2}{a vector of shrunken McFadden's pseudo R^2 measures for \code{r} LFs.}
}
}
\author{
Neo Chung \email{nchchung@gmail.com}
}
\references{
Chung and Storey (2015) Forthcoming
}
\seealso{
\link{jackstraw.LFA} \link{jaws.pca}
}

