% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaws_pca.R
\name{jaws.pca}
\alias{jaws.pca}
\title{The Jackstraw Weighted Shrinkage Estimation Method for Sparse Loadings in Principal Component Analysis}
\usage{
jaws.pca(dat, p = NULL, r = NULL, s = NULL, B = NULL,
  stat.shrinkage = "F-statistics", extra.shrinkage = NULL, verbose = TRUE,
  seed = NULL, save.all = TRUE)
}
\arguments{
\item{dat}{a data matrix with \code{m} rows as variables and \code{n} columns as observations.}

\item{p}{a \code{m * r} matrix of p-values for association tests between variables and \code{r} principal components, generally computed from the jackstraw method. If \code{p} is not given, \code{jackstraw.PCA} is automatically applied.}

\item{r}{a number (a positive integer) of significance principal components.}

\item{s}{a number (a positive integer) of ``synthetic'' null variables (optional).}

\item{B}{a number (a positive integer) of resampling iterations (optional).}

\item{stat.shrinkage}{PNV shrinkage may be applied to "F-statistics" or "loadings" (default: F-statistics).}

\item{extra.shrinkage}{extra shrinkage methods may be used; see details below (optional).}

\item{verbose}{a logical specifying to print the progress (default: TRUE).}

\item{seed}{a seed for the random number generator (optional).}

\item{save.all}{a logical specifying to save all objects, including a large SVD object (default: FALSE).}
}
\value{
\code{jaws.pca} returns a list consisting of
\item{p}{p-values for association tests between variables and each of \code{r} principal components}
\item{pi0}{proportion of variables not associated with \code{r} principal components, individually}
\item{svd}{SVD object from decomposing \code{dat}}
\item{PIP}{a list of outputs derived from the posterior inclusion probabilities method (including \code{pr}, \code{u}, \code{var}, \code{PVE})}
\item{PNV}{a list of outputs derived from the proportion of null variables method (including \code{pi0}, \code{u}, \code{var}, \code{PVE})}

With appropriate \code{extra.shrinkage} options (for details, see the Supplementary Information of Chung and Storey (2013), the output may also include
\item{PIPhard}{a list of outputs from hard-threshoding the \code{PIP} loadings (including \code{u}, \code{var}, \code{PVE})}
\item{PIPsoft}{a list of outputs from soft-threshoding the \code{PIP} loadings (including \code{pr}, \code{u}, \code{var}, \code{PVE})}
}
\description{
Estimates sparse/shrunken loadings of principal component analysis.
Based on statistical sginificance of association between variables and principal components,
the sample loadings of principal components are shruken towards zeros, which improve its accuracy.
The only required inputs are the data matrix \code{dat} and the number of principal components \code{r} whose loadings you would like to estimate.
}
\details{
By default, \code{jaws.pca} computes two canonical jackstraw weighted shrinkage estimators, namely \code{PIP} and \code{PNV}.
Additionally, other extra shrinkage techniques may apply, such as combining two canonical estimaotrs by setting \code{extra.shrinkage="PIPhard"}
and applying soft-thresholding to local fdr by setting \code{extra.shrinkage} to numerical threshold values between 0 and 1.
Please provide \code{r} numerical threshold values to be applied to \code{r} principal components.

It is strongly advised that you take a careful look at your data and use appropriate graphical and statistical criteria
to determine a number of significant PCs, \code{r}. For example, see a contributed R package called `nFactors'.
In a case when you fail to specify \code{r}, \code{r} will be estimated from permutation Parallel Analysis (Buja and Eyuboglu, 1992)
via a function \link{permutationPA}, with a very liberal threshold.

If \code{s} is not supplied, \code{s} is set to about 10\% of \code{m} variables.
If \code{B} is not supplied, \code{B} is set to \code{m*10/s}.
}
\section{Detailed explanation of the output contained in \code{PIP} and \code{PNV}}{

\describe{
  \item{pr}{a matrix of posterior inclusion probabilities (equivalent to 1-lfdr) for \code{m} coefficients and \code{r} PCs.}
  \item{pi0}{a vector of estimated proportion of null variables for \code{r} PCs.}
  \item{u}{a \code{m*r} matrix of shrunken loadings.}
  \item{var}{a vector of shrunken variances explained by \code{r} PCs.}
  \item{PVE}{a vector of shrunken percent variances explained by \code{r} PCs.}
}
}
\examples{
set.seed(1234)
## simulate data from a latent variable model: Y = BX + E
B = c(rep(1,50),rep(-1,50), rep(0,900))
X = rnorm(20)
E = matrix(rnorm(1000*20), nrow=1000)
dat = B \%*\% t(X) + E
dat = t(scale(t(dat), center=TRUE, scale=FALSE))

## estimate sparse loadings in PCA
jaws.pca.out = jaws.pca(dat, r=1)

}
\author{
Neo Chung \email{nchchung@gmail.com}
}
\references{
Chung and Storey (2015) Forthcoming
}
\seealso{
\link{jackstraw.PCA} \link{jaws.cov}
}

